unsigned int

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: scharuka <scharuka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/03 00:18:13 by scharuka          #+#    #+#             */
-/*   Updated: 2022/08/03 00:23:58 by scharuka         ###   ########.fr       */
+/*   Updated: 2022/08/03 00:24:59 by scharuka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = ft_strlen(s);
 	while (i >= 0)
